Backend/Django/requirements.txt
Django==3.1.14
djangorestframework==3.13.1
Faker==12.0.0
PyJWT==2.3.0
mysqlclient==2.1.0
django-cors-headers
Pillow==9.0.1
